requirements.txt
pandas>=2.0.0
numpy>=1.24.0
httpx>=0.25.0
web3>=6.0.0
pydantic>=2.0.0
pyyaml>=6.0
sqlalchemy>=2.0.0
scipy>=1.10.0
